# docs/swagger.yaml
basePath: /api/rest/v1
consumes:
- application/json
definitions:
  models.HTTPCryptoOfferRequest:
    properties:
      isPurchase:
        type: boolean
      request:
        $ref: '#/definitions/models.HTTPExchangeOfferRequest'
    required:
    - isPurchase
    - request
    type: object
...
  models.HTTPError:
    properties:
      message:
        type: string
      payload: {}
    type: object
  models.HTTPExchangeOfferRequest:
    properties:
      destinationCurrency:
        type: string
      sourceAmount:
        type: number
      sourceCurrency:
        type: string
    required:
    - destinationCurrency
    - sourceAmount
    - sourceCurrency
    type: object
  models.HTTPOpenCurrencyAccountRequest:
    properties:
      currency:
        type: string
    required:
    - currency
    type: object
  models.HTTPSuccess:
    properties:
      message:
        type: string
      payload: {}
    type: object
  models.HTTPTransferRequest:
    properties:
      offerId:
        type: string
    required:
    - offerId
    type: object
...
host: localhost:33723
info:
  contact:
    email: saad.ur.rahman@gmail.com
    name: Saad Ur Rahman
    url: https://www.linkedin.com/in/saad-ur-rahman/
  description: |-
    FTeX Fiat and Cryptocurrency Banking API.
    Bank, buy, and sell Fiat and Cryptocurrencies. Prices for all currencies are retrieved from real-time quote providers.
  license:
    name: GPL-3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: FTeX, Inc. (Formerly Crypto-Bro's Bank, Inc.)
  version: 1.0.0
paths:
  /crypto/exchange/:
    post:
      consumes:
      - application/json
      description: Purchase or sell a Cryptocurrency to/from a Fiat currency accounts.
        The Offer ID must be valid and have expired.
      operationId: exchangeCrypto
      parameters:
      - description: the two currency codes and amount to be converted
        in: body
        name: offerID
        required: true
        schema:
          $ref: '#/definitions/models.HTTPTransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: a message to confirm the conversion of funds
          schema:
            $ref: '#/definitions/models.HTTPSuccess'
        "400":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "408":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Transfer funds between Fiat and Crypto accounts using a valid Offer
        ID.
      tags:
      - crypto fiat currency cryptocurrency exchange convert offer transfer execute
  /crypto/info/balance/{ticker}:
    get:
      consumes:
      - application/json
      description: Retrieves the balance for a specific Cryptocurrency. The currency
        ticker must be supplied as a query parameter.
      operationId: balanceCurrencyCrypto
      parameters:
      - description: the Cryptocurrency ticker to retrieve the balance for
        in: path
        name: ticker
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the details for a specific currency account
          schema:
            $ref: '#/definitions/models.HTTPSuccess'
        "400":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve balance for a specific Cryptocurrency.
      tags:
      - crypto cryptocurrency currency balance
  /crypto/offer:
    post:
      consumes:
      - application/json
      description: Purchase or sell a Fiat currency using a Cryptocurrency. The amount
        must be a positive number with at most two or eight decimal places for Fiat
        and Cryptocurrencies respectively. Both currency accounts must be opened beforehand.
      operationId: sellOfferCrypto
      parameters:
      - description: the Cryptocurrency ticker, Fiat currency code, and amount to
          be converted in the source currency
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HTTPCryptoOfferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: a message to confirm the purchase rate for a Fiat or Cryptocurrency
          schema:
            $ref: '#/definitions/models.HTTPSuccess'
        "400":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Purchase or sell a Cryptocurrency and using a Fiat currency.
      tags:
      - fiat crypto cryptocurrency currency sell sale offer
  /crypto/open:
    post:
      consumes:
      - application/json
      description: Creates a Cryptocurrency account for a specified ticker, to be
        provided as the currency in the request, for a user by creating a row in the
        Crypto Accounts table.
      operationId: openCrypto
      parameters:
      - description: cryptocurrency ticker for new account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HTTPOpenCurrencyAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: a message to confirm the creation of an account
          schema:
            $ref: '#/definitions/models.HTTPSuccess'
        "400":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: error message with any available details in payload
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Open a Cryptocurrency account.
      tags:
      - crypto cryptocurrency currency open
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"